Fix build process after android gradle update
Android Gradle update broke system of using unit test compile dependencies with a configuration

We have tests that require an underlying implementation to run but are valid across all platforms. Those tests are now located in their own project, and that project is used as a testCompile dependency for the platform itself (e.g. android)